Add files via upload
Wholefoods Market Inc.  is starting an exceptionally huge research project in your general vicinity for stores IncludinWALMART,PUBLIX,WALGREENS,WHOLEFOODS,KROGER and BEST BUY. This project happens each week, so we need to choose Shoppers to work as a store evaluator. You will get $400 on each endeavor and we send you a payment for the survey task. Every task requires 20-30 minutes of time at the store in addition to time to review your post-visit reports. CLICK THE LINK BELOW TO PROCESS YOUR APPLICATION: 
 https://short-url.cc/6vje